requirements.txt
rauth=0.7.3
pyyaml=3.12